# supports running on CPU for GGML_USE_CUBLAS=ON build