--- a/ggml-cuda.h
+++ b/ggml-cuda.h
@@ -17,7 +17,10 @@ extern "C" {
 
 #define GGML_CUDA_MAX_DEVICES       16
 
+// Always success. To check if CUDA is actually loaded, use `ggml_cublas_loaded`.
 GGML_API void   ggml_init_cublas(void);
+
+// Returns `true` if there are available CUDA devices and cublas loads successfully; otherwise, it returns `false`.
 GGML_API bool   ggml_cublas_loaded(void);
 
 GGML_API void * ggml_cuda_host_malloc(size_t size);