@ggerganov 2023-11-05
We might want to merge something like this PR, so that 3rd party projects have an easier way to support optional CPU-only runs. Though, I'm not sure if this is the best way to do.
Any alternative suggestions?