@wsxiaoys 2023-11-05
> We might want to merge something like this PR, so that 3rd party projects have an easier way to support optional CPU-only runs. Though, I'm not sure if this is the best way to do. Any alternative suggestions?

Some alternative ideas were considered when prototyping this pull request (assuming `GGML_USE_CUBLAS=ON`):

1. Utilizing the `CUDA_VISIBLE_DEVICES` environment variable: If no valid indices are present in `CUDA_VISIBLE_DEVICES`, and there is insufficient deduction for CUDA availability, we can expose an API that is valid without depending on `ggml_cublas_init`.

2. Relocating `ggml_cublas_init` to `llama.cpp`, similar to the 'metal' backend: I haven't delved deeply into this approach, as it appears to be incompatible with the ggml backend v2.

Please let me know if you have any questions or need further clarification on these alternatives.