@slaren 2023-11-05
Most of this will become obsolete after llama.cpp is adapted to use ggml-backend. After that, the way this will be implemented by making `ggml_backend_cuda_init` return `NULL` (or something to the same effect). I cannot say when ggml-backend will be ready to be used in llama.cpp, so I think it is ok to add a temporary solution, but considering that, the best solution IMO is whatever is easiest to remove in the future.