@wsxiaoys 2023-11-06
I cleaned up it a bit and think it should be relative easy for a ggml_backend migration.