@wsxiaoys 2023-11-06
> Note that simply setting -ngl 0 would not work because ggml will keep moving data to the device and do some of the computations there instead of the CPU

I experimented with `-ngl 0`, and it appeared to match the performance of the CPU-only option. Nevertheless, this may not be a very common use case. Therefore, we should consider recommending the usage of `CUDA_VISIBLE_DEVICES=-1` instead.