@ggerganov 2023-11-06
It's possible that the behaviour with `-ngl 0` has changed and now would actually perform everything on the CPU without moving data back and forth as it used to do in the past. If you don't see any GPU usage in your test, then all should be good.