@zhangp365 2023-11-07
I compiled the code using the latest version and copied the resulting executable file to a computer without a graphics card and no CUDA installed. When I attempted to run the main.exe file, it failed to execute. The relevant error message is as follows:
![error](https://github.com/ggerganov/llama.cpp/assets/144313702/02fbdb8e-0727-4a66-9f66-81c7f998065b)