@wsxiaoys 2024-02-03
Due to newer changes adding link against libcuda - the fix is no longer working.

It will generates following error message for a cuda build when running in non-cuda environment:

```
/opt/tabby/bin/tabby: error while loading shared libraries: libcuda.so.1: cannot open shared object file: No such file or directory
```

Looking into workaround